update: totalAmount

+++ serivce_request/src/main/java/com/companies/serivce_request/service/ServiceRequestService.java
@@ -9,6 +9,7 @@ import com.companies.serivce_request.repository.ServiceRequestRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -57,7 +58,7 @@ public class ServiceRequestService {
         entity.setInspectionRequired(Boolean.TRUE.equals(svc.defaultInspectionRequired()));
 
         // Fees Amount
-        entity.setTotalFeesAmount(req.totalFeesAmount());
+        entity.setTotalFeesAmount(BigDecimal.valueOf(req.totalFeesAmount()));
         entity.setFeesApproved(true);
 
         // Owning agency as default primary agency
@@ -162,7 +163,7 @@ public class ServiceRequestService {
                 .orElseThrow(() -> new ResourceNotFoundException("Service request not found: " + id));
 
         if (req.totalFeesAmount() != null) {
-            entity.setTotalFeesAmount(req.totalFeesAmount());
+            entity.setTotalFeesAmount(BigDecimal.valueOf(req.totalFeesAmount()));
         }
         if (req.serviceFeesJson() != null) {
             entity.setServiceFeesJson(req.serviceFeesJson());
@@ -232,7 +233,7 @@ public class ServiceRequestService {
                 e.isInspectionRequired(),
                 e.getInspectionStatus(),
                 e.getInspectionReference(),
-                e.getTotalFeesAmount(),
+                e.getTotalFeesAmount().doubleValue(),
                 e.getServiceFeesJson(),
                 e.getFeesApproved(),
                 e.getPaymentStatus(),
